requirements.txt
# ==========================================
# Manga Reader API - Requirements (FULL FEATURES)
# ==========================================
# Version: 4.0.0
# Python: 3.9+
# Features: Reading, Bookmarks, Lists, Analytics, Upload
# ==========================================

# Core Framework
# ==========================================
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.1.0
python-multipart==0.0.9

# Database & ORM
# ==========================================
sqlalchemy==2.0.27
pymysql==1.1.0
cryptography==42.0.2
alembic==1.13.1

# Authentication & Security
# ==========================================
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyjwt==2.8.0
psutil==5.9.6
# Background Tasks & Scheduling
# ==========================================
apscheduler==3.10.4

# Environment & Configuration
# ==========================================
python-dotenv==1.0.1

# HTTP Client
# ==========================================
httpx==0.26.0
requests==2.31.0

# Logging & Monitoring
# ==========================================
python-json-logger==2.0.7
structlog==24.1.0
coloredlogs==15.0.1

# Rate Limiting
# ==========================================
slowapi==0.1.9
limits==3.10.1

# Redis Cache (Optional)
# ==========================================
redis==5.0.1
hiredis==2.3.2

# Monitoring (Optional)
# ==========================================
sentry-sdk[fastapi]==1.40.4
prometheus-fastapi-instrumentator==7.0.0
prometheus-client==0.19.0

# Testing
# ==========================================
pytest==7.4.4
pytest-asyncio==0.23.4
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
# ==========================================
black==24.1.1
flake8==7.0.0
mypy==1.8.0
isort==5.13.2
pylint==3.0.3
httpx==0.27.0
# Input Sanitization
# ==========================================
bleach==6.1.0

# Performance
# ==========================================
orjson==3.9.13
ujson==5.9.0

# Utilities
# ==========================================
python-slugify==8.0.1
validators==0.22.0

# Image Processing
# ==========================================
pillow==10.2.0
python-magic==0.4.27
aiofiles==23.2.1
filetype==1.2.0


email-validator==2.3.0
dnspython==2.8.0
